content-governance-suite/requirements.txt
# Core API dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

# AI/ML dependencies - Flexible versions
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
diffusers>=0.20.0
accelerate>=0.20.0
transformers>=4.30.0

# LangChain - Use compatible ranges
langchain>=0.1.0,<1.0.0
langchain-core>=0.1.0,<1.0.0
langchain-community>=0.0.20,<1.0.0
langchain-perplexity>=0.1.0

# Other ML dependencies
sentence-transformers>=2.0.0
opencv-python>=4.5.0
pillow>=9.0.0
imageio-ffmpeg>=0.4.0

# Core dependencies
python-dotenv>=0.19.0
numpy>=1.21.0
pandas>=1.3.0

# Database - Flexible versions
sqlalchemy>=1.4.0,<3.0.0
psycopg2-binary>=2.8.0
redis>=4.0.0

# Export features
reportlab>=3.6.0
python-docx>=0.8.11

# Additional
pydantic>=1.8.0
aiofiles>=0.8.0
